PID实验

+++ MyProject_test/PID/PID.h
@@ -6,14 +6,14 @@ typedef struct PID{
 	uint16_t Sv; //设定的目标值
 	uint16_t Ep; //当前的测量值
 	
-	int16_t Kp; //当前参数
+	float Kp; //当前参数
 	float Ti; //积分常量
 	float Td; //微分常量
-	int16_t T;  //PID控制周期
+	float T;  //PID控制周期
 	
-	int16_t Ek; //当前误差
-	int16_t Ek_1; //上一次的误差
-	int32_t SEk; //历史误差
+	float Ek; //当前误差
+	float Ek_1; //上一次的误差
+	float SEk; //历史误差
 	
 	float Pout; //比例输出
 	float Iout; //积分输出
